vehicle-spec-extractor/requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-huggingface
faiss-cpu
pymupdf
pandas
python-dotenv
tiktoken
sentence-transformers
